switch amount and balance to bigdecimal

+++ src/main/java/com/accenture/multibank/accounts/AccountModifiable.java
@@ -1,8 +1,10 @@
 package com.accenture.multibank.accounts;
 
+import java.math.BigDecimal;
+
 public interface AccountModifiable extends AccountReadable {
 
-	AccountReadable book(int amount);
+	AccountReadable book(BigDecimal amount);
 
 
 }
