+++ src/main/java/com/accenture/multibank/bank/Bank.java
@@ -1,8 +1,11 @@
 package com.accenture.multibank.bank;
 
+import java.math.BigDecimal;
+
+import org.springframework.stereotype.Service;
+
 import com.accenture.multibank.accounts.AccountReadable;
 import com.accenture.multibank.accounts.AccountType;
-import org.springframework.stereotype.Service;
 
 /**
  * @author manuel
@@ -11,8 +14,11 @@ import org.springframework.stereotype.Service;
 @Service
 public interface Bank {
     // TODO: Generic bank because the account number may not be represented by an integer
-    AccountReadable withdraw(Integer accountNumber, int amount);
-    AccountReadable deposit(Integer accountNumber, int amount);
-    AccountReadable transfer(Integer fromAccountNumber, Integer toAccountNumber, int amount);
-    AccountReadable createAccount(AccountType type, int balance);
+	AccountReadable withdraw(Integer accountNumber, BigDecimal amount);
+
+	AccountReadable deposit(Integer accountNumber, BigDecimal amount);
+
+	AccountReadable transfer(Integer fromAccountNumber, Integer toAccountNumber, BigDecimal amount);
+
+	AccountReadable createAccount(AccountType type, BigDecimal balance);
 }
